requirements.txt
asgiref==3.11.0
Django==5.0.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==12.1.0
requests==2.32.3
reportlab==4.2.5
sqlparse==0.5.5
